backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
langchain==0.0.286
langchain-experimental==0.0.16
langgraph==0.0.15
openai==1.3.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
elasticsearch==8.9.0
redis==4.6.0
stripe==6.0.0 